refactor(dashboard): move layout helpers out of shared modules
- Move mergeLayout from DashboardMixin to src/pages/dashboardLayout.ts
- Move getStoredPanels from the gui store getters to a PageDashboard method
- Move the mergeLayout unit tests to tests/pages/dashboardLayout.spec.ts
